inline variable: apply inline variable for totalVolumeCredits
Now the author suggests to apply the inline variable in order to inline
totalVolumeCredits. In this refactoring he is always trying to get less
variables as possibles since it's believed that by doing that temporary
variables won't pollute the code and any other refactoring become easier
without variables around the codebase that would need to be taken into
account, e.g. when extracting methods.